requirements.txt
aiogram==2.23.1
python_dotenv==0.0.5
mysql-connector-python==9.0.0
